pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autoskillit"
version = "0.5.2"
description = "MCP server for orchestrating automated skill-driven workflows with Claude Code"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Talon Trecek", email = "talontrecek@gmail.com"},
]
keywords = ["claude", "claude-code", "automation", "ai", "mcp", "plugin"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.6.0",
    "dynaconf>=3.2.12,<4.0",
    "fastmcp>=3.1.1,<4.0",
    "httpx>=0.28,<0.29",
    "igraph>=1.0",
    "packaging>=23.0",
    "psutil>=7.2.0",
    "pyyaml>=6.0",
    "cyclopts>=4.0",
    "structlog>=25.0,<26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",  # asyncio_default_fixture_loop_scope requires >=1.0.0
    "pytest-httpx>=0.36",
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.3",
    "ruff>=0.15.0",
    "import-linter>=2.1",
    "packaging>=25.0",
]

[project.urls]
Homepage = "https://github.com/TalonT-Org/AutoSkillit"
Repository = "https://github.com/TalonT-Org/AutoSkillit"
Documentation = "https://github.com/TalonT-Org/AutoSkillit/tree/stable/docs"
"Bug Tracker" = "https://github.com/TalonT-Org/AutoSkillit/issues"

[project.scripts]
autoskillit = "autoskillit.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/autoskillit"]
artifacts = [
    "src/autoskillit/.claude-plugin/**",
    "src/autoskillit/.mcp.json",
    "src/autoskillit/migrations/**",
    "src/autoskillit/hooks/**",
    "src/autoskillit/config/defaults.yaml",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/autoskillit/**",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
empty_parameter_set_mark = "fail_at_collect"
addopts = []
timeout = 60
timeout_method = "signal"
markers = [
    "smoke: end-to-end smoke tests requiring Claude API access (deselect with: -m 'not smoke')",
]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "TID251"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"logging.getLogger" = {msg = "Use get_logger() from autoskillit._logging instead."}

[tool.importlinter]
root_packages = ["autoskillit"]
# Exclude imports made inside `if TYPE_CHECKING:` guards from all contracts.
# This is critical for execution/headless.py, which imports pipeline types only
# inside TYPE_CHECKING blocks to avoid circular imports at runtime.
exclude_type_checking_imports = true

# IL-001 | Lens: module-dependency | DS-001
# Rationale: L0 (core) is the foundation layer; it must have zero upward imports to
# remain independently testable and importable outside the autoskillit package.
[[tool.importlinter.contracts]]
name = "L0 core has no autoskillit upward imports"
type = "forbidden"
source_modules = ["autoskillit.core"]
forbidden_modules = [
    "autoskillit.config",
    "autoskillit.pipeline",
    "autoskillit.execution",
    "autoskillit.workspace",
    "autoskillit.recipe",
    "autoskillit.migration",
    "autoskillit.server",
    "autoskillit.cli",
]

# IL-002 | Lens: module-dependency | DS-001
# Rationale: L1 config resolves settings from layered sources; allowing it to import
# pipeline/execution/recipe would create hidden initialization-time coupling that
# breaks independent layered testing.
[[tool.importlinter.contracts]]
name = "L1 config imports only L0"
type = "forbidden"
source_modules = ["autoskillit.config"]
forbidden_modules = [
    "autoskillit.pipeline",
    "autoskillit.execution",
    "autoskillit.workspace",
    "autoskillit.recipe",
    "autoskillit.migration",
    "autoskillit.server",
    "autoskillit.cli",
]

# IL-003 | Lens: module-dependency | DS-001
# Rationale: pipeline/ holds cross-session gate and audit state; importing from L2
# (recipe, migration) would create initialization-order coupling and prevent L1
# packages from being tested in isolation without a full L2 import graph.
[[tool.importlinter.contracts]]
name = "L1 pipeline does not import L2 or L3"
type = "forbidden"
source_modules = ["autoskillit.pipeline"]
forbidden_modules = [
    "autoskillit.execution",
    "autoskillit.workspace",
    "autoskillit.recipe",
    "autoskillit.migration",
    "autoskillit.server",
    "autoskillit.cli",
]

# IL-004 | Lens: module-dependency | DS-001
# Rationale: execution/ manages subprocess lifecycle; it must not import config,
# pipeline, workspace, recipe, migration, server, or cli to remain a leaf L1 service
# that can be tested without spinning up the full autoskillit stack.
[[tool.importlinter.contracts]]
name = "L1 execution imports only L0"
type = "forbidden"
source_modules = ["autoskillit.execution"]
forbidden_modules = [
    "autoskillit.config",
    "autoskillit.pipeline",
    "autoskillit.workspace",
    "autoskillit.recipe",
    "autoskillit.migration",
    "autoskillit.server",
    "autoskillit.cli",
]

# IL-005 | Lens: module-dependency | DS-001
# Rationale: workspace/ manages cloning and skill resolution; importing from config
# or pipeline would introduce a circular risk since config and pipeline both may
# depend on workspace for path resolution.
[[tool.importlinter.contracts]]
name = "L1 workspace imports only L0"
type = "forbidden"
source_modules = ["autoskillit.workspace"]
forbidden_modules = [
    "autoskillit.config",
    "autoskillit.pipeline",
    "autoskillit.execution",
    "autoskillit.recipe",
    "autoskillit.migration",
    "autoskillit.server",
    "autoskillit.cli",
]

# IL-006 | Lens: module-dependency | DS-001
# Rationale: recipe/ is an L2 service that orchestrates validation over workspace
# artifacts; importing config, pipeline, execution, migration, server, or cli
# would conflate recipe semantics with runtime session management.
[[tool.importlinter.contracts]]
name = "L2 recipe imports only L0 and workspace"
type = "forbidden"
source_modules = ["autoskillit.recipe"]
forbidden_modules = [
    "autoskillit.config",
    "autoskillit.pipeline",
    "autoskillit.execution",
    "autoskillit.migration",
    "autoskillit.server",
    "autoskillit.cli",
]

# IL-007 | Lens: module-dependency | DS-001
# Rationale: migration/ must not import config or pipeline to stay independently
# runnable as a schema-upgrade step without a live pipeline context.
[[tool.importlinter.contracts]]
name = "L2 migration does not import L3 or config"
type = "forbidden"
source_modules = ["autoskillit.migration"]
forbidden_modules = [
    "autoskillit.config",
    "autoskillit.pipeline",
    "autoskillit.server",
    "autoskillit.cli",
]
